Manage PhaseThree better using a state machine
Since I can't just shove all the initialization stuff in the ready
function, I realized we were going to need some pretty complex logic
to handle what needs to happen where and when.

I was sort of bouncing around this idea of multiple scenes that passed
around common data (like the BlockCache) when I remembered any
complex-logic's best friend: The State Machine!

After perusing the options, I went with imjp94's Yet Another Finite
State Machine (YAFSM). The simplicity matched up with the needs of
this project (it's not my insanity RTS, after all).

I then adjusted the PhaseThree scene to utilize the state machine.
Everything worked well, save for a few caveats: I had to maintain the
one-second pause when we first enter the scene, AND I had to increase
the block-build-gap to .5 seconds. I'm not sure what caused that gap
increase, to be honest. It just started crashing on its own!