Organize blocks into a faux-city
I originally created the CopyBlock with a two-node idea: the CopyBlock
which would display the mesh, and some sort of placeholder that didn't
have anything BUT would replace itself with the CopyBlock on demand.

Over time, I realized there was no added value in two separate scenes,
so I overhauled the CopyBlock into the FacsimilieBlock, which acts as
both placeholder and actual physical object.

I then created the organizational level beyond this - the city. I
eventually plan on there being multiple kinds of city. However, for
now, we just have the LinearCity. Now I need to figure out a way to
make the LinearCity truly functional...