Build dynamically using feedback from the blocks
So we got the blocks into the city; the next step was to somehow start
placing them dynamically.

After bouncing around a bunch of different ideas, I came up with one
that worked pretty well: the blocks would dynamically place down
blocks to their left or right when they spawned in. This allows us to
create a chain that extends, potentially infinitely, parallel to a
particular axis.

I'm pretty happy with how it works. I just wish there was an obvious
way to make all that happen in two directions...