Clean up neighbors when a block leaves the scene
Okay, we used to scroll infinitely, and generate infinite buildings.
Fun, except for the fact that we don't have infinite memory. So, we
needed to start cleaning up the buildings once they left the scene.

Since we spawned in buildings whenever they entered the scene, it felt
like the most logical point to delete them was when they left the
scene.

Except, we didn't want to do it RIGHT when they left the scene - after
all, we rely on seed blocks that sit off screen and spawn more blocks
once on screen. After some coniving, I came up with a solution: if a
block moves offscreen, and it's neighbor was already offscreen, then
that is deleted. It's like the chain-creation, but instead it's chain
deletion.

Honestly I'm not crazy about this solution because some creeping
monster in my mind implanted back in my college days KNOWS there some
hole here that can VERY EASILY go wrong. But it probably won't and
this whole thing is a giant hack so.... meh.