Move block-cleanup to the city level
Previously, we had the blocks clean up their own neighbors when they
went offscreen. This was a fine enough idea, but I started to have
ideas about the city having more command-and-control over the blocks.
In order to do that, the city would need to control removal of the
blocks to the same degree that it controls the block creation.
Besides, seperating block creation-and-deletion is just bad
programming practice.

So, I moved the block-deletion code up into the LinearCity code. I had
to make use of the group call since signals are inherently
one-source-to-many functions, instead of many sources-to-one-function.