Assign block groups on a per-city basis
Previously, the facsimilie blocks all belonged to a generic group -
"blocks". We just used that to touch all the blocks at once. But that
wasn't good enough. This project supports (on paper at least) two
seperate types of city, and thus two simultaneous existing set of
blocks.

Treating all the blocks as a monolithic group would thusly have
ramifications on all of the city blocks. THAT WOULD NOT DO!

So, we now assign the blocks to a city-specific group at block
creation. That will make things neat and orderly if we ever get around
to actually doing that orthographic city...