Use MATH to ensure blocks are placed on camera
Previously, we were sort of vaguely guessing where to place the
LinearCity so that the seed block just happened to appear on camera.
That was good enough for one row, but not good enough for multiple
rows. Since the camera is SUPPOSED to be at an angle, we can't dummy
spawn the seed blocks in a series either.

No, I'm afraid we needed to use HIGH SCHOOL GEOMETRY to spawn those
rows.

So, I added multiple control variables for the number of rows as well
as the initial offset and extra distance between block rows. I then
added functions to calculate the intricate positions - I left a bevy
of comments so I think I can skip the logic behind that.

The result is we now generate three seperate rows of blocks at
runtime. Success! Now I either need to tackle the loadtimes (they're
driving me insane) or I need to handle building selection...