Add a block gluer prototype that parses lines
Okay, I'm getting annoyed at my own project's load times. Too many
buildings, all that waiting. It's too inefficient. But I can't load
the individual blocks any faster! I need to take everything and sort
of... MASH IT TOGETHER.

I've already edited a Qodot file in Godot. It should be easy to glue
together something via a script. Actually, maybe I could do that with
Godot at runtime? You know what, let's just do it in Python first. The
pseudocode before the real deal!

To start, I noticed that the structure of the TrenchBroom map file
consists of entities and brushes contained there in. The structure's
actually pretty easy to parse: a '{' indicates that we're entering an
entity or a brush. A '}' indicates that we're leaving that entity or
brush. Plus, each one of those is contained on it's own line!

So, the current glue script just goes moves between three different
states depending on the lines encountered. Easy!