Process face lines into an object & spit them out
Okay, so a Quake map is only composed of entities and brush
definitions. Entities contain brushes, and brushes are composed of
faces.

That means the most basal level of processing we can do is to process
the face-definition lines that compose a block. Thus, I added code to
do JUST that!

Although there were several ways we could have gone about it, I
decided that the BEST way was to try and approximate GodotScript. I
was tempted to make specific classes for Brushes and BrushFaces, but
GodotScript doesn't allow for such custom class definitions (I think).

So, INSTEAD, I created two classes to approximate Godot's Vector2 and
Vector3 types. What was going to be the BrushFace object is now a
function and a patchwork dictionary. Sloppy, in some ways, but it's
good pseudo-code for our Godot implementation.