Add the ability to shift Vectors
I added the ability to "shift" my faux GD Vectors while maintaining
the precision of the floating point values. It's something Python CAN
do but generally elects not to because 90% of the time it DOESN'T
MATTER.

Except since we're trying to just modify an existing value with an
existing precision, it does matter. The solution is some wacky string
bending that feels heretical but that's because I'm trying to emulate
the behavior of one language in an entirely seperate language.

Eh, whatever. Now to get some REAL functionality in here.