Process multiple maps via a function and arguments
If we're going to glue multiple maps together, then we're going to
need to process multiple maps. It's only natural!

So, I moved the whole map-processing loop into a function. We now
assume that all the arguments passed into the script are map paths.
We need at least one, or we back out. Pass into the function to get
entities, and then...

Well, right now we just run a count of the different brushes and
entities we have. Actually doing something is the NEXT step.