Translate parsed maps into groups and print-to-file
All the information was available to us; we now just needed to sort of
stitch it together and save it to a file.

In order for everything to work correctly, we needed to group the
blocks together using a func_group. That allows Qodot to build up a
particular block as a mesh. I tested it out and it seems to work fine.
Now we need to move this thing into Godot. That'll be the real test...