Set up func_group; track properties per-block
If we're going to save time by mashing all of the Qodot blocks
together into some sort of mushy Frankenstein, we're going to need a
way to differentiate between all those different Frankenstein parts
so we can easily pick out blocks and other features from the block
cache.

To do that, we needed to somehow group together the disparate point
entities. Best way to do that? It'd actually be placing the nodes
together in a tree structure, except Qodot doesn't do that. So,
instead, we'll assign everything to a Godot group as it comes in.

Thus, I modified the func_group to take in and store some extra
properties as well as set it's group at run time. I did the same for
the light entity/class as well. It seems to be working, for now. We'll
need to rewrite how we copy things from the BlockCache, but this is
definitely the best way forward right now.