Qodot Hotfix
While I'm mostly done with this project (for now at least), I was
trying to test performance on an old college laptop when I found that
Qodot wasn't working! Turns out there's different Qodot versions for
different versions of Godot, which... makes sense. My bad, I dropped
the ball there.

Anyway, I've added in the updated files. This is kind of slapped
together (you're supposed to delete the qodot folder entirely) but
that's all I'm willing to commit to this project at this time.