Create footprints by growing points outward
I previously tried creating footprints by growing out Voronoi polygons
and... I don't know, I didn't care for it. It seemed to me we needed a
more square footprint. Or rectangular footprint. Just, not whatever
the Voronoi Polygons were offering up.

I worked for about a week on a system where we randomly placed points
along the z axis and then formed boxes by connecting points as AABBs.
That system was bad. Or maybe bad so much as "generally not good". It
glitched a lot and had a LOT of code.

Eventually, I realized I could streamline things by seeding points and
then growing the AABBs (or rectangles, as most people call them)
outward. I just needed to stop them when they broke out of our z limit
or collided with each other.

This new system works out pretty good, just so long as we don't place
the points too close together.

Now the current demonstrator is just that - a demonstrator. We need to
find a way to do this progressively, always spawning things on the
right. And we need to do the efficiently, too...