Give points a gaussian distribution
Okay, so I was trying to figure out how to actually plant/distribute
the points. I knew that I wanted there to be more points at the top,
less points towards the bottom.

And then I realized it - what I wanted was a Gaussian distribution!
So, I've started using a gaussian distribution to determine placement
on Z - it's generated with a mean of zero, and a preconfigured
deviance. We take the absolute value and then place the points at the
lower limit. The total effect is that we there's a greater density of
GrowPoints placed toward the Lower Z limit, and a lesser density of
GrowPoints towards the Upper Z limit.

The Max Size of a GrowPoint is a square, determined by checking the Z
point against a curve. This allows us to generate larger buildings at
the back, and smaller buildings towards the front. Perfect!

We also now stepify our points, which helps clean some stuff up.