Isolate GrowAABB in it's own script
The ultimate idea for this growth point nonsense is like so: we plant
a bunch of growth points all around a certain area, which we'll most
likely term a block. We spawn these blocks in, and they are managed by
some sort of manager that we'll call GrowthCity, or something. Or
GrowBlockifier. Or something. Anyway, GrowBlockifier will spawn the
blocks and handle the point growths (possibly in some sort of thread).
It will offer up completed GrowAABB objects to whoever wants them:
ideally, these will then be translated into building footprints, and
new buildings. At least, that's the hope.

Anyway, the first step to making all this happen is to try and isolate
the GrowAABB. So, I did that!

I did try converting it to work with the UV-size calculations that I
made - the GlobalRef WINDOW_UV_SIZE constant. However, this meant we
were comparing floats and that started to have some weird side
effects. So, I skipped it. We'll assume that the AABB lengths are
measured in number-of-windows, since we only use whole numbers.