Create a GrowBlockifier to manage GrowAABBs
Now that GrowAABBs are in their own script, we need a script/object
to manage the growth of these objects.

I've implemented this as the GrowBlockifier - it basically just takes
the guts of the GrowPointTest and isolates it in a BlockTest. We don't
have any logic for why we spawn blocks, or really where we spawn
blocks - but spawn blocks we do!

Now, let's try working with more than one block at a time. The
blockifier does have to spawn blocks, after all! Then we can take a
look at making the GrowBlockifier threaded in some way.