Manage three blocks at a time, building them progressively
Doing just one block at a time is fine, but we're going for big game
here! Block after block after block!

To do that, we need the Blockifier to manage our blocks for us. So,
we have two new management functions - grow_all_blocks() to grow
everything, and clean_and_spawn_pass() to clean up old blocks and
add new blocks.

One of the issues with the grow points is that they change depending
on the presence of their neighbors - if they hit a neighbor, they
stop. That's good functionality, we want to keep that. That just means
that if there are no neighbors, buildings can end up looking weird.

However, if everything's growing at the same rate, when do you start
trimming off the blocks? You'd have to sort of figure out the exact
amount of blocks you need, such that you're adding new blocks at the
rate you're deleting them. And that's imprecise, hard to calculate,
SPECIFICALLY because the organic growth algorithm that helps each
city look unique.

After giving it some thought, I think I could skip that whole block
add-and-delete problem by making the blocks grow AGGRESSIVELY in a
progressive fashion.

It's kind hacky, but it works like this: we currently have three
blocks. When we do a grow pass on the blocks, we grow the first block
three times, the second block two times, and the last block one time.
This makes our farthest-left block, which is theoretically the
most-required block, the fastest growing. It routinely finishes before
the middle block, which finishes before the right-hand block. Ergo, we
can ensure that we can complete blocks faster than we're generating
them. That means we can generate at-leisure, instead of trying to
figure out the timing!