Organize spawning, growing, and cleaning into passes
Okay, we had the growing and spawning functionality set up. We just
needed to get some cleaning in place, as well as get everything more
organized. I made that last commit when I was tired, so there was some
stuff to polish up around the edges.

So, we now have the three phases of a block lifecycle tied to
different "pass" functions: spawn_pass, grow_pass, and clean_pass. Now
clean pass is the only one worth talking, because it's the only truly
new one.

We 'clean' (delete) a block when both it, and it's neighbor to the
right, are no longer viable. Initially it was both left-and-right
neighbor, but I realized the infinite right-side growth meant that we
only needed to car about our right-hand neighbor. If a given block is
finished growing, then it follows that the block to it's left is
already done growing. So, no need to track the left neighbor - only
the right-hand neighbor!

I also upped the number of blocks to 4 (so we have
edge-regular-regular-edge). I separated the progressive grow count
into a variable separate from the block count. Finally, I noticed that
some grow points were growing into each other because they were both
being placed on the edges of blocks - to compensate, we now add a
shift of blank space between blocks.