Uncap positive growth on Z
I realized that, for the purposes of this project, the Z-plane is
basically infinite. So I uncapped the placement of points on Z and the
growth of AABBs on +Z. Unfortunately that shot up the amount of time
growing was taking (I think some points were getting placed very far
off screen) so I capped placement, but not growth. This produces more
of the "super structures" that I had dreamed about.

I did notice a few weird things though - one, our timer kept
automatically restarting even though it wasn't configured to - I had
to purposefully stop it with stop().

The other odd thing was that, while cleaning up a block, the value
returned by a binary search went out-of-bounds on one of the
directional AABB sorts. I noticed that there was about a 50 point
difference between some of the arrays! I need to get to the bottom of
that...