Sort for block changes; avoid binary search's inaccuracies
We were getting this weird phenonmenon where the custom binary search
was returning the wrong index when searching a sorted array. This
meant that we weren't properly cleaning up any of our AABBs from our
sorted arrays. And THAT was leaving us with some unnecessary search
overhead.

I couldn't figure out why, though. I did realize we were never
re-sorting the arrays after we grew in a particular direction, so they
were technically unsorted. That was an improvement, but it didn't fix
the issue.

I realized after a time, that the bsearch_custom function was just
straight returning the wrong index - and I think I know why. It
expects a function that returns true if a < b. I THINK, because of the
way the compare functions were designed, it just stops when it hits
the first x or z value that is equal.

That works for boundary checking, but not actually getting our index
to erase ourselves. So, we now use the manual erase() method, which is
slower but actually works. In addition, I updated the boundary-checking
algorithm: we now either return BEFORE or AFTER the given value, in
order to maximize the AABBs we can check against.