Boil GrowPointTest into GrowthTest and BlockTest
After all this time, I've finally managed to build the grow
point-and-block algorithm into something that works pretty reliably.

So, it was time to translate that into actual buildings. I broke the
GrowPointTest into the GrowthTest and made a copy, the BlockTest. The
BlockTest is like the GrowthTest, but tests translating the blocks
into actual buildings.

I also removed the mesh stuff from GrowAABB, because that class
doesn't actually need any visual representation. It was mostly a
conveinence for the GrowPointTest. Of course, that meant I had to
change the GrowthTest to generate and update these meshes so I could
actually look at them. But, all seems to be working now.