Properly instantiate the Blockifier
Previously, we would create the blockifier and then manually update
the Blockifier's variables - length, width, curve, etc. I've updated
the Blockifier with an init() function, so we now pass these in to
the Blockifier when it's created. The Blockifier now also takes in the
number of points to spawn in.

I've also been playing around with different cameras for the
BlockTest. It's all looking pretty good so far!