Vary the heights of buildings
The city wouldn't look very good if all of the buildings were the same
height, would it? We now use minimum and maximum height curves to
determine the range of a building's possible height. The new-style
city is already looking better, though it honestly looks almost
indistinguishable from the old-style city.