Add a building that randomizes itself to a footprint
Previous iterations of the this project relied on the
FullServiceBuilding, which generates a building to a given height,
width, length, etc.

Now this is good, but we need to do something better. See, one of the
big problems we've had with the previous iterations is how they're
somewhat plain looking. So we need the ability to work within a
specific sort of space - a footprint - and generate things within
those specific confinies.

That's what the new FootprintBuilding does. You set the footprint
boundaries, and the height of the tower, and it should take care of
the rest. Right now, it only moves the building around - but we should
be able to add some lights here...