Add in a light-based shader
We worked with adding in lights to our city before. I tried just using
lights, and I also tried a weird dither-grid thing. It looked kinda bad
so I stopped using it.

However, we're trying to add more depth to the city now. And one of the
best ways to do that would be with lighting. So, I've reimplemented
that earlier lighting algorithm on a copy of the DotWindowShader,
creating the brand new Dot Window Light Shader.

So, it works pretty good, though it needs at least one light source in
order to appear at all. To get around that, I placed a directional light
in the BlockTest facing upward. Bit hacky, bit stupid, but it works!