Spawn a light for each FootprintBuilding
We're going to try using lights for our new buildings - and, to do
that, we're going to need to actually spawn in some lights.

So, I cooked up a whole complicated mess capable of testing
whether-or-not a point was intersecting with the rotated rectangle.
The idea was that we'd use this to avoid putting lights in the rotated
rectangle.

Just for now, I put the light in one corner of the footprint. And then
I realized - why put it anywhere other than the footprint? To be
honest, I can't very well think of a reason. I'll have to ponder on
that...