Make the new grow-block style the primary city
I've removed the old styl of city generation and replaced it with the
GrowBlockCity, which is basically a growth test made exclusively for
growing. Now, I'm not gonna lie, it doesn't run that great, but that's
okay - we can do some threadery and hopefully get rid of any nonsense.

Honestly, it's a bit broken right now. The camera slowly drifts into
the buildings and for some reason the building node kept flipping
itself to -180. But, it works! It's a start, at least.