Cap the amount of shrink-adjustments for FootprintBuildings
Sometimes the city was freezing - a quick stack trace revealed that we
were shrinking the building infinitely. And that just won't do! So, I
capped the amount of rolls we can do for a building. If we exceed the
acceptable rolls, we just default to a standard orientation. Not
particularly exciting, but it works.