Have buildings remove themselves when exiting the screen
I was able to identify one major source of our woes: none of the
buildings were actually deleting themselves. Oops. I think I had
intended for to clean up the buildings when their parent block was
deleted.

However, I observed that since this way of deleting affects the
lights, actual "visibility" extends beyond the borders of a block.
So it's better, then, to have each building manage itself.

So each building now deletes itself once it exits the screen. I've
added an auto clean option to turn this auto-deleting off.
Eventually we'll need to scale the visibility for the radius of the
lights we spawn...