Thread-ize placing and designing buildings
We were getting REALLY framey as time went on. Just massive pauses and
other such hiccups.

I was able to determine that we were losing our framerate mostly to
the omnilights casting shadows. So unfortunately I had to get rid of
that.

However, that still left us with large lag spikes. Observing the
frame profiler, I found the culprits: growing the buildings, and
randomly shrinking/rotating the buildings. Honestly, was that even
surprising? I had sort of deliberately designed this system to be
somewhat inefficient. I knew what I was doing going in. I had
purposefully set it up so that, if we needed to, we could thread the
intensive stuff - and we needed to. These grow-and-spawn frames were
taking upwards of 200 ms.

So, I threaded the GrowBlockCity's growing sequence, and the
FootprintBuilding's construction sequence. Now you can't spawn in
nodes from a thread, so we not everything is threaded. What we did
instead was thread the resource intensive calculations that we were
doing - the ones that told us what to do with a node.

Anyway, it seems to be working excellently now. Though it does start
to slow down from 110 FPS to ~90 FPS as the simulation drags on...