Adjust main scene for Grow Block City
I needed to make some adjustments to the main city scene because the
GrowBlockCity scene was so different. Moving the camera, moving the
starfield, adjusting the size of the city. So on and so on.

Now I've noticed that spawning an appropriately scene-filling building
count (~40 buildings per block) is causing us frame drops. I need to
find a way to not build all of the buildings at once...