Only make buildings to keep under a certain frame duration
We were getting big bad frame processing spikes whenvever we spawned
in new buildings. It wasn't as bad as our pre-thread state, but it
still wasn't great.

The problem was that we were starting our blueprinting threads all at
once, then finishing them all at once, then constructing all of the
buildings at once. Again, not as bad a spike as we could be getting,
but still not great.

So, I've changed how we make FootprintBuildings once again. We now
emit a signal once the blueprinting thread is complete. From that
signal, we queue up the building in a to-be-built array. When we time
to process the array, we build as many buildings as possible while
trying to stay under the millisecond target.

The result is better, but not perfect. I messed around with some
settings and got things looking a little bit better. Honestly, it
seems like setting the physics framerate higher brings us greater
success. The issue is that physics framerate eats massively into our
CPU. I haven't noticed any OS problems spinning off from that, but
being a CPU hog isn't something I'm entirely comfortable with...