Add some TODO statements to help guide the project
I've decided that, given the visual nature of this project, the sudden
FPS dips are unacceptable. Even if not running live, the FPS dips are
noticeable in recorded footage. I can't allow that.

That means we can't do an infinite city - but we can do a very large
city that loops! That would allow this program to produce a smoothly
looped video or possibly even a GIF.

I left a whole host of TODOs to help us chart our direction for this
rework. This ain't gonna be easy...