Remove YAFSM
I included the YAFSM add-on because I wanted a light-weight state
machine to help smooth out some of the more complex processes. I've
used a different finite state machine in one of my other projects
before - the XSM. XSM was heavily code-based, though, so I wanted to
see if I could try a more visual alternative: the YAFSM.

Unfortunately I never actually got it to work. Actually, I might have
gotten it to work once but it was such a pain I took it out. If I had
one criticism/complaint, it's that the YAFSM was too passive. It had
to be given values and signals in order to transition between states.
This is different from XSM, where you manually triggered the state
transitions directly.

Ultimately, the system was too fiddly for me to bother using it. This
project is cluttered enough in it's fifth-or-so iteration, so I'm
happy to delete it.