Show & hide the FootprintBuilding when it enters & leaves
Previously, we would have a FootprintBuilding remove itself from a
scene and then delete itself. Good for an infinite city, less good for
a generated city of a fixed size. Plus, that level of self-management
was largely conceived of as a lazy shortcut, and not as some act of
good design. Givng a node free-reign to create and destroy itself
seemed like a bad idea.

So, we now automatically show and hide the visually-intense components
of a building when it enters and exits a scene, respectively. We also
emit signals out whenever the building enters and exits the scene, so
that whatever is managing the building can handle that.