Add a function to launch a blueprint make thread
I added a multi-threaded blueprint build process to the
FootprintBuilding in the absolutely laziest way possible - it only
launched the threaded blueprint make if you enabled auto build and the
building entered the scene.

I like my code to be a bit more feature-complete than that. So, I
added a function that launches a blueprint make thread that we can
call from anywhere. It won't launch another thread if there's one
currently active.