Generate FootprintBuildings using a seed
In order to generate buildings consistently, we need to pull the same
random values. And for that, we need a seed. So, I added an init()
function that takes in a string-seed. If there's no string seed, it'll
roll itself a random seed.

The idea is we'll hand it a string that is the same for a particular
building - something like the building's origin and it's size - along
with something like the city's seed. The same city will generate the
same footprints which will generate the same buildings!