Generate a city of a fixed size
Previously, we generated the city infinitely. That was pretty easy to
manage; but as we've stated, this caused unacceptable performance
issues.

The next best thing to 'infinite' is 'seemingly infinite'. In that
case, we need a city that grows to fixed degree and then loops almost
seamlessly.

I'm not going to lie, that was painful to implement. I took a lot of
shortcuts making the city infinite so actually capping the stupid
thing took longer than it needed to.

We now keep all the blocks around, but selectively ignore the ones
that have been cleaned. To complete the effect of a loop, we copy the
first block to the end so the ACTUAL last block grows and collides
with the first block. However, we set it up so that the fake last
block is never actually spawned.

The performance is pretty good. About as well as can be expected, I
suppose. You can generate up to 100 blocks, but I find that 50 is more
than sufficient.

I did have to change how we updated the sorted arrays: previously, we
were just calling the binary_sort() function on the whole array. That
was fine when we had four blocks tops; it got a lot messier once we
had upwards of 2000 buildings. So we now have to manually find,
remove, re-search, and re-insert the updated AABB. That's also
intensive, as you can imagine, but it's better. Only takes like
10-some-odd seconds.