Wrap buildings around once they go off-screen
Previously, our loop was incomplete because buildings were going
offscreen and then never coming back onscreen. I've changed it so
that whenever a building goes offscreen, we now warp it to the other
side of the city - completing the loop!

I also noticed that I wasn't really applying the window scaling to the
height of the building - that meant I had to adjust all of our height
curves to get everything fixed.