Scale buildings in Grow Block City using a scalar
I was admittedly lazy when it came to scaling the buildings: I scaled
up the BuildingMaster node, which was parent to all of the buildings,
up. That automatically scaled all the buildings and made my life easy.

Unfortunately, we need the buildings to actually know how big they are
and adjust their omnilights appropriately. That meant scaling the
buildings individually. No more just slapping stuff under nodes and
calling it a day.

Fortunately, the actual scaling part wasn't all that difficult. Our
code is a bit gnarlier now that we're putting in the scalar constant,
but otherwise isn't worse-for-wear.