Incorporate omnilight into AABB calculation
Previously, we weren't incorporating the omnilight radius of a
particular building into the actual AABB. This meant that the building
was being removed too early, when it technically was still affecting
the scene. That was causing lighting updates, breaking the illusion of
a continous city.

We couldn't do that before, partially because I was lazy but also
partially because the of the lights themselves. They need to be scaled
manually since that don't scale with a node's "scale" field. Since the
BuildingManager was the one doing that for use, we couldn't access
that info and scale appropriately.

But, now that the scale is being handed down to the building, we can
scale the light and see how far we're supposed to go! Though I figured
out that - since our light is measured in window lengths - it doesn't
really make a difference.

Anyway, instead of smartly calculating the light's position and where
to stretch the AABB, I decided to assume the a light somehow exists on
all four corners. Lazy, but easy, and effective!

I was also able to observe and correct another bug, where buildings
weren't being move appropriately because I forgot to include the
BUILDING_SCALAR.