Isolate the building tab into it's own scene
Up until now our GUI has just been this sort of giant terrible monster
script with a whole mess of code about updating the various materials
and fixtures that make up the Main scene.

I've decided that, to try and simplify things, we're going to split
each tab off into it's own scene that will be exclusively dedicated to
it's one... thing. These individual tabs will then interact with a
common dictionary that we can save and read from a JSON file.

So I started with the building tab, which seemed like a good first
choice to me. It offered us a lot of opportunities to clean up the
Main script, though in truth that thing is still a mess.

Utilizing what I learned from the Qodot arc, I changed the building
tab to dynamically populate the choices menu using our window assets.
Eventually, we'll be able to spin this into loading user textures.

We now also "assert" the status of a common dict from the material
color control autoload node. We used to do this in main, but will now
do it from a common and omnipresent source. This is closer to my
original design intention, which I deviated from for... some reason.