Update the Block Test
We changed the Blockifier a whole lot, but never updated the Block
Test to match. This update fixes that - we removed the bake button,
and actually make the buildings in a threaded context.

I was fixing to remove the Voronoi colors, but I decided to use them
for this test - the multicolor buildings actually look pretty nice.

Plus, they help us differentiate between overlapping buildings.