Isolate Moon and Starfield Tabs; adjust MCC piecemeal
Okay, so I thought I had this whole tab isolation nonsense down-pat,
so I charged ahead and isolated the Moon, Starfield, and Star tabs.
However, I ran into such big problems with the Moon and Starfield tabs
that I had to backpedal on the Star Tab.

The biggest problem I had was that the profile was getting
overwritten. After some work, I found that this was because of the
various update_to_global() functions being called from the
update_from_global() functions, which was happening because we were
updating the GUI. Unrelated, but similar, we were starting to get
major frame drops when changing stuff that used to be easy to change.

The solution was two-fold: first, we disable updating a profile value
when we're reading from the profile. Secondly, instead of updating all
the values at once, we now update the MCC materials using an
update_key() function. This allows us to update JUST what we need.

This is working really well. Now to isolate those stars.