Isolate the Stars Tab
The next tab to isolate was the Stars tab. This one was a bit more
complex as, like the Building Tab, we needed to dynamically load in
some textures. However, as we've already done most of the heavy
lifting in the Building Tab, it was easy to just copy the pattern for
the Stars Tab.