Add a profile tab
After all of that nonsense, the last thing we had to do was update the
Profile tab. Of course, the profile tab needed actual profiles to
load.

Now previously all the profiles were defined, by hand, as
dictionaries. Obviously that wasn't very reproducable, so I endeavored
to store all of the profiles as JSON files and load them as assets so
we'd easily have them on hand.

EXCEPT the Godot editor doesn't recognize JSON files! Straight-up
refuses to display them. So that wasn't an option. I ultimately
figured out I could use a custom resource with an easy-convert function
to produce a dictionary.

The profile tab just loads an array of those resources, which then
becomes the profile selection. Choosing a profile changes everything
appropriately. We don't have a clear execution order for profile
changes. Everything's kind of muddled between the Main class, the MCC,
and the tabs. We'll need to do something about that.

Also, as a side note, the MoonTab wasn't updating the moon's
visibility properly. It was waiting for the button to be pressed, not
for the value to change. This has been updated accordingly.