Clarify the relationship between city, MCC, and GUI
Okay so before the MCC was just used for changing the materials but
also storing the profile BUT the profile was updated by the GUI and
then the GUI needed to also update just certain parts of the city
and... If you can't tell from the run-on sentence, nothing really made
sense.

So, I sat down and created a clearer pathway. The GUI updates the
profile in the MCC, which then takes the appropriate material-control
(i.e. global/project level) actions. It then echoes out a signal,
which any listening scene can then use to update the values
appropriately.

The only place where this nice flow gets weird is the GUI, because the
GUI needs to update itself to reflect changes within the profile. So
the GUI can literally update the profile and then immediately receive
an update from the MCC for the change it just made! We do have things
in place to stop an infinite loop, and it doesn't seem to impact
performance too much, so we can get away with it. FOR NOW!

The other hiccup was that we technically have a profile when we start,
so we need to do an initial update. To sort of force that, I added the
assert_profile() function. This just propagates the current profile
(i.e. the default) around to the GUI, the MCC, and the city.