Add an option to rotate the sky
You could see more of the sky in the old city design. I like the
distribution of buildings in this city, but my only lament was that
you don't get to see as much of the sky. Some of those color profiles
I cooked up looked wonderful against the sky. Really added to the
effect, and the newer system was so building rich you couldn't see the
sky anymore.

Then I found out you can rotate the sky using the Environment
resource. That fixes it! I've added the sky's X rotation as a
configurable under the Sky tab (the Y needs to be fixed to the camera
angle).

The only thing that I noticed was that the sky doesn't like certain
angles. Everything starts getting flashy plus-or-minus 100 degrees. No
idea what's going on there!